+++ scrooge-generator-tests/src/test/resources/gold_file_output_java/com/twitter/scrooge/test/gold/thriftjava/CollectionId.java
@@ -26,6 +26,7 @@ import org.apache.thrift.protocol.*;
 
 import com.twitter.scrooge.ThriftStructIface;
 import com.twitter.scrooge.TFieldBlob;
+import com.twitter.scrooge.internal.TProtocols;
 
 // No additional import required for struct/union.
 
@@ -340,12 +341,9 @@ public class CollectionId implements TBase<CollectionId, CollectionId._Fields>, 
       }
       switch (field.id) {
         case 1: // COLLECTION_LONG_ID
-          if (field.type == TType.I64) {
-            this.collectionLongId = iprot.readI64();
-            setCollectionLongIdIsSet(true);
-          } else {
-            TProtocolUtil.skip(iprot, field.type);
-          }
+          TProtocols.validateFieldType(TType.I64, field.type, "collectionLongId");
+          this.collectionLongId = iprot.readI64();
+          setCollectionLongIdIsSet(true);
           break;
         default:
           if (this.passThroughFields == null) {